Move sources into src folder Add tests folder with live test sandbox view Use scss for styles Split polyfills into separate files
