+++ package.json
@@ -4,9 +4,10 @@
   "description": "JavaScript segments percentages slier",
   "main": "package/index.js",
   "scripts": {
-    "start": "node_modules/.bin/parcel ./index.html",
-    "build-prod": "node_modules/.bin/parcel build -d ./build --no-source-maps ./index.html --public-url ./",
-    "build-package": "node_modules/.bin/parcel build -d ./package --no-content-hash --no-source-maps --out-file percentage-slider ./index.ts",
+    "start": "npm run live-test",
+    "live-test": "node_modules/.bin/parcel ./src/tests/live/index.html",
+    "build-prod": "node_modules/.bin/parcel build -d ./build --no-source-maps ./src/tests/live/index.html --public-url ./",
+    "build-package": "node_modules/.bin/parcel build -d ./package --no-content-hash --no-source-maps --out-file percentage-slider ./src/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -27,6 +28,7 @@
   "homepage": "https://github.com/Zarwlar/percentage-slider#readme",
   "devDependencies": {
     "parcel": "^1.12.4",
+    "sass": "^1.32.8",
     "tslib": "^1.10.0",
     "typescript": "^3.6.4"
   }
