+++ src/polyfills/Array.ts
@@ -1,0 +1,25 @@
+// Array.prototype.findIndex
+(function () {
+  if (!Array.prototype.findIndex) {
+    Array.prototype.findIndex = function (predicate) {
+      if (this == null) {
+        throw new TypeError('Array.prototype.findIndex called on null or undefined');
+      }
+      if (typeof predicate !== 'function') {
+        throw new TypeError('predicate must be a function');
+      }
+      let list = Object(this);
+      let length = list.length >>> 0;
+      let thisArg = arguments[1];
+      let value;
+
+      for (let i = 0; i < length; i++) {
+        value = list[i];
+        if (predicate.call(thisArg, value, i, list)) {
+          return i;
+        }
+      }
+      return -1;
+    };
+  }
+})();
