+++ src/tests/README.md
@@ -1,0 +1,2 @@
+No any automatic tests for now: only the live example
+that could be run by `npm run live-test`.
