+++ src/tests/live/index.html
@@ -7,8 +7,6 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <meta name="description" content="JavaScript percentage slider. Demo page.">
   <title>Document</title>
-  <link rel="stylesheet" type="text/css" href="./percentage-slider.css">
-  <link rel="stylesheet" type="text/css" href="./client.css">
 </head>
 
 <body>
